+++ src/pages/Bus/BusCoursePage/components/BusTimetable/BusTimetable.module.scss
@@ -1,0 +1,70 @@
+@use "src/utils/scss/media" as media;
+
+.template {
+  width: 795px;
+  margin: 23px auto 0;
+
+  @include media.media-breakpoint(mobile) {
+    min-height: 350px;
+    width: 100%;
+    margin: 0;
+    box-sizing: border-box;
+  }
+
+  &__title {
+    font-size: 20px;
+    font-weight: bold;
+    letter-spacing: -1px;
+    color: #252525;
+    margin-bottom: 22px;
+  }
+
+  &__loading-spinner {
+    margin: 250px auto;
+    background-color: white;
+    width: 100px;
+    height: 100px;
+  }
+}
+
+.tabs {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  height: 56px;
+  list-style: none;
+
+  @include media.media-breakpoint(mobile) {
+    width: 100vw;
+    height: 100%;
+  }
+
+  &__tab {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 16px;
+    letter-spacing: -0.8px;
+    font-weight: 500;
+    line-height: 1.5;
+    color: rgb(133 133 133);
+    background-color: #fff;
+    appearance: none;
+    border-radius: 12px 0 0;
+    font-family: Pretendard, sans-serif;
+    border-bottom: 2px solid #cacaca;
+    padding: 16px 16px 12px;
+    cursor: pointer;
+
+    &--selected {
+      color: rgb(23 92 142);
+      border-bottom: 3px solid rgb(23 92 142);
+    }
+
+    @include media.media-breakpoint(mobile) {
+      display: flex;
+      font-size: 16px;
+      padding: 12px 16px;
+    }
+  }
+}
